Final version functional and increase the delay function bits

+++ Delay.h
@@ -10,6 +10,6 @@
 
 #include <stdint.h>
 
-void delay(uint16_t delay);
+void delay(uint32_t delay);
 
 #endif /* DELAY_H_ */
